requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
adafruit-circuitpython-neopixel>=6.3.7
pydantic>=2.0.0
pydantic-extra-types>=2.0.0